Bug fixes in unit tests and code cleanup

+++ Application/UseCases/Users/DeleteUserUseCase.cs
@@ -14,6 +14,14 @@ public class DeleteUserUseCase : IDeleteUserUseCase
 
     public async Task Execute(string id)
     {
+        if (id == null)
+        {
+            throw new ArgumentNullException(nameof(id), "Id cannot be null");
+        }
+        if (id == string.Empty)
+        {
+            throw new ArgumentException("Id cannot be empty", nameof(id));
+        }
         await _userRepository.DeleteUser(id);
     }
 }
